learning: DRobotPerceptron: Support multiple learning rules
In addition to Oja's rule, provide the learning rule used by McMillen et
al. 2011 (or in a similar form by Cook et al. 2010) and the Hebbian
covariance rule (Sejnowski et al. in R. Durbin et al., The Computing
Neuron, 1989)

The learning rule to use can be chosen by means of passing a parameter
of type learn_rule_t to the updateWeights(...) functions.

Also provide a winner-take-all weight-update function which will
dertermine the max. rate neuron and only learn weights to it (and
neighbors).